requirements.txt
PySimpleSOAP==1.05a
phpserialize==1.3
python-dateutil==2.1
